Rename test; Attempt to use custom container for testing env vars
